IS-3348: Endret fra Spek til JUnit
