+++ build.gradle.kts
@@ -1,8 +1,9 @@
+import com.adarshr.gradle.testlogger.theme.ThemeType
+
 group = "no.nav.syfo"
 version = "0.0.1"
 
 val jacksonDataType = "2.18.3"
-val kluent = "1.73"
 val ktor = "3.1.2"
 val logback = "1.5.17"
 val logbackSyslog4jVersion = "1.0.0"
@@ -10,13 +11,13 @@ val logstashEncoder = "7.4"
 val micrometerRegistry = "1.12.13"
 val mockk = "1.13.17"
 val nimbusJoseJwt = "10.2"
-val spek = "2.0.19"
 val jedis = "5.2.0"
 
 plugins {
     kotlin("jvm") version "2.1.10"
     id("com.gradleup.shadow") version "8.3.6"
     id("org.jlleitschuh.gradle.ktlint") version "11.4.1"
+    id("com.adarshr.test-logger") version "4.0.0"
 }
 
 repositories {
@@ -57,9 +58,7 @@ dependencies {
     testImplementation("io.mockk:mockk:$mockk")
     testImplementation("io.ktor:ktor-client-mock:$ktor")
     testImplementation("com.nimbusds:nimbus-jose-jwt:$nimbusJoseJwt")
-    testImplementation("org.amshove.kluent:kluent:$kluent")
-    testImplementation("org.spekframework.spek2:spek-dsl-jvm:$spek")
-    testRuntimeOnly("org.spekframework.spek2:spek-runner-junit5:$spek")
+    testImplementation(kotlin("test"))
 }
 
 kotlin {
@@ -84,9 +83,11 @@ tasks {
     }
 
     test {
-        useJUnitPlatform {
-            includeEngines("spek2")
+        useJUnitPlatform()
+        testlogger {
+            theme = ThemeType.STANDARD_PARALLEL
+            showFullStackTraces = true
+            showPassed = false
         }
-        testLogging.showStandardStreams = true
     }
 }
